backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymupdf>=1.23.0
python-docx>=0.8.11
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
werkzeug>=3.0.0
gunicorn>=21.0.0